+++ src/js/videojs.record.js
@@ -1456,10 +1456,10 @@
                 case this.ANIMATION:
                     this.streamCurrentTime = Math.min(currentTime, duration);
 
-                    // update control
-                    this.player().controlBar.currentTimeDisplay.el(
-                        ).firstChild.innerHTML = this.formatTime(
-                        this.streamCurrentTime, duration);
+                    // update current time display component
+                    this.player().controlBar.currentTimeDisplay.formattedTime_ = this.player(
+                        ).controlBar.currentTimeDisplay.contentEl(
+                        ).lastChild.textContent = this.formatTime(this.streamCurrentTime, duration);
                     break;
             }
         },
@@ -1495,10 +1495,10 @@
                 case this.VIDEO_ONLY:
                 case this.AUDIO_VIDEO:
                 case this.ANIMATION:
-                    // update control
-                    this.player().controlBar.durationDisplay.el(
-                        ).firstChild.innerHTML = this.formatTime(
-                        duration, duration);
+                    // update duration display component
+                    this.player().controlBar.durationDisplay.formattedTime_ = this.player(
+                        ).controlBar.durationDisplay.contentEl(
+                        ).lastChild.textContent = this.formatTime(duration, duration);
                     break;
             }
         },